pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mermaid-mcp"
version = "0.1.0"
description = "基于MCP协议的流程图生成服务器"
authors = [
    {name = "Chasey-AI", email = "chasey.agent@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
dependencies = [
    "mcp>=1.8.0",
    "fastapi>=0.95.0",
    "uvicorn>=0.22.0",
    "playwright>=1.32.0",
    "openai>=1.0.0",
    "anthropic>=0.5.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "jinja2>=3.1.2"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.3.1",
]

[tool.setuptools]
packages = ["src"] 